Move tests to /tests/it/array/dictionary on request. Satisfy Miri with doctest
